Use int32_t to pass data, which can avoid data overflow when shifting towards left.

+++ singleStagePipeline/src/Memory.cpp
@@ -4,7 +4,7 @@ void Memory::Add(const int32_t address, const uint8_t data) {
   memories[address] = data;
 }
 
-uint8_t Memory::Get(const int32_t address) {
+int32_t Memory::Get(const int32_t address) {
   if (this->IsEmpty(address)) {
     memories[address] = 0x00;
   }
